fix includes for ct_slavery add a simple unit test for charming effects
