+++ src/economy.c
@@ -22,6 +22,7 @@
 #include "spells/buildingcurse.h"
 #include "spells/regioncurse.h"
 #include "spells/unitcurse.h"
+#include "spells/charming.h"
 
 /* kernel includes */
 #include "kernel/ally.h"
