+++ src/spells/charming.h
@@ -1,5 +1,9 @@
 #pragma once
 
 struct castorder;
+struct unit;
 
 int sp_charmingsong(struct castorder *co);
+void charm_unit(struct unit *target, struct unit *mage, double force, int duration);
+
+extern const struct curse_type ct_slavery;
